+++ src/gui/fasto_common_item.cpp
@@ -76,7 +76,7 @@ QString toJson(FastoCommonItem* item) {
   }
 
   QString value;
-  for (int i = 0; i < item->childrenCount(); ++i) {
+  for (size_t i = 0; i < item->childrenCount(); ++i) {
     value += toJson(dynamic_cast<FastoCommonItem*>(item->child(i)));
   }
